test(bb-agent): fold runaway-cap regression tests into index.test.ts (no separate files)
Moves all three cap regression tests into the existing 'runaway protection
caps' suite in index.test.ts and deletes the two standalone files, so the
package keeps a single unit-test entry point and the tests run under the
default test script without listing extra files:

- maxLlmCalls keeps counting after resume() (fails if the counters are
  per-segment locals)
- an approved tool call is charged once, not twice, across resume()
- a second message on the same conversation gets a fresh budget (fails against
  an eager, pre-session-restore reset)

The shared runAutoApproving() helper moves alongside them. No assertions were
weakened. The package test script is back to
'node --test dist/index.test.js && node --conditions=cdk --test dist/index.cdk.test.js'.

bb-agent unit: 87 pass / 0 fail; cdk: 3 pass / 0 fail; lint clean.
